Added Spring Security

+++ src/main/java/service/ParcelService.java
@@ -1,7 +1,7 @@
 package service;
 
-import dao.AwaitListDao;
-import dao.ParcelDao;
+
+import dao.*;
 import entity.AwaitList;
 import entity.Parcel;
 import entity.User;
@@ -9,19 +9,24 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.Date;
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
-public class ParcelService {
+public class ParcelService implements IParcelService {
 
-    private ParcelDao parcelDao;
+    private IParcelDao parcelDao;
 
-    private AwaitListDao awaitListDao;
+    private IWaiters awaitListDao;
+
+    private IUserDao userDao;
 
     @Autowired
-    public ParcelService(ParcelDao parcelDao, AwaitListDao awaitListDao) {
+    public ParcelService(IParcelDao parcelDao, IWaiters awaitListDao,IUserDao userDao) {
         this.parcelDao = parcelDao;
         this.awaitListDao = awaitListDao;
+        this.userDao=userDao;
     }
 
     @Transactional
@@ -33,21 +38,31 @@ public class ParcelService {
     public void registerParcelByRegcode(String regcode, User recipient){
 
         Parcel parcel=parcelDao.getParcelByRecipientAndRegcode(recipient,regcode);
-        parcel.setStatus("Recived");
-        awaitListDao.addUserToAwaitList(parcel.getMailer());
+        if(parcel!=null) {
+            parcel.setStatus("Received");
+            awaitListDao.addUserToAwaitList(parcel.getMailer());
+        }
     }
 
     @Transactional
     public Parcel addNewParcel(User mailer){
         AwaitList awaitList=awaitListDao.getFirstRecordFromAwaitList(mailer);
-        User recipient = awaitList.getUser();
-        awaitListDao.delete(awaitList);
+        User recipient;
+        if (awaitList==null){
+            recipient = userDao.getRandomUserButNotCurrent(mailer);
+        }else {
+            recipient= awaitList.getUser();
+            awaitListDao.delete(awaitList);
+        }
+        LocalDate sendDate= LocalDate.now();
+
         Parcel parcel = new Parcel();
+        parcel.setSendDate(Date.valueOf(sendDate));
         parcel.setStatus("Sended");
         parcel.setMailer(mailer);
         parcel.setRecipient(recipient);
-        String mailerCountry = mailer.getAddress().getCountryName();
-        String recipientCountry = recipient.getAddress().getCountryName();
+        String mailerCountry = mailer.getAddress().getCountry().getCountryShortcut();
+        String recipientCountry = recipient.getAddress().getCountry().getCountryShortcut();
         parcel.setRegistrationCode(generateRegcode(mailerCountry,recipientCountry));
         parcelDao.add(parcel);
         return parcel;
@@ -59,13 +74,42 @@ public class ParcelService {
     }
 
     @Transactional
-    public List<Parcel> getAllRecivedByUser(User user){
-        return parcelDao.getAllRecivedUserParcels(user);
+    public List<Parcel> getAllReceivedByUser(User user){
+        return parcelDao.getAllReceivedParcelsByUser(user);
     }
 
     private String generateRegcode(String mailerCountry,String recipientCountry){
         String regcode = (mailerCountry+(int)(Math.random() * 1000000) + recipientCountry);
         return regcode;
+    }
+
+    @Transactional
+    public List<Parcel> getLastParsels(final int COUNT_OF_PARCELS_TO_DISPLAY){
+        List<Parcel> parcels=getAll();
+        if(parcels.size()>COUNT_OF_PARCELS_TO_DISPLAY){
+            parcels =parcels.subList(parcels.size()-COUNT_OF_PARCELS_TO_DISPLAY,parcels.size());}
+        return parcels;
+    }
+
+    @Transactional
+    public List<Parcel> getLastSendedUserParsels(final int COUNT_OF_PARCELS_TO_DISPLAY,User user){
+        List<Parcel> parcels=getAllSendedByUser(user);
+        if(parcels!=null) {
+            if (parcels.size() > COUNT_OF_PARCELS_TO_DISPLAY) {
+                parcels = parcels.subList(parcels.size() - COUNT_OF_PARCELS_TO_DISPLAY, parcels.size());
+            }
+        }
+        return parcels;
+    }
+
+    @Transactional
+    public int getCountOfParcels(){
+       return parcelDao.getCountOfParcels();
+    }
+
+    @Transactional
+    public int getCountOfReceivedParcels(){
+        return parcelDao.getCountOfReceivedParcels();
     }
